We define the Kron radius as the major axis a, not sqrt(ab)
I think that we should change to using sqrt(ab), but that's a
different changeset -- I want to get all the tests passing first so we're
back to where we were first.

N.b. We might want to put out the Kron moments instead for consistency with other
measurements.  At present we save R_Kron and a/b and phi are implicit
in the use of getShape() to define them.