calculatePsfKronRadius: use PSF radius instead of calculating ourselves
RHL writes: """
I think we should just use the 2nd moment. The comment complains that we
can't use computeGaussianWidth, but it's been superseded by
computeShape().getDeterminantRadius() which is const-qualified. If we do
this the quadrature sum of the moment and smoothingSigma is correct, and
the sqrt(pi/2) should go outside the sqrt.
"""