add more options for setting the aperture shape
Previously, we have given up when provided with a bad shape for the source,
but we can set the shape to something sensible in that case, like the
shape of the PSF.

This removes responsibility for choosing the aperture shape from
KronAperture::determine and places it in KronFlux::_apply, where all the
information is available (e.g., the Control object).  We introduce a new
flag, flux.kron.flags.badShape, that indicates the input shape was bad and
we used the PSF shape instead.