add fallback options for setting Kron radius
The Kron radius, being the integral of the first moment of the radius over
the footprint, doesn't behave well at low signal-to-noise ratio.  In order
to get a value out of the Kron algorithm, we now fall back to using the
Kron radius of the PSF (unless a minimumRadius is set, in which case we
use that instead).

Introduced two new flags (usedMinimumRadius and usedPsfRadius) to indicate
what actual radius was used.