only set algorithm master flag if there was a fundamental problem
We want to set the algorithm master flag if there's a problem that renders
the measurement of low quality.  We don't want to set the algorithm master
flag if the problem is just low signal-to-noise, in which case using the
PSF Kron radius is probably just fine for getting a decent measurement.